Move check-in code generator to models module
